relocate coverage/efficiency/average_interval_width to probly.metrics
Make probly.metrics the canonical home for all scoring functions. The
three predicted-set metrics now live alongside auc, roc_curve, etc. and
are re-exported from probly.evaluation so existing callers (six examples,
benchmark, all docstrings) keep working unchanged.

- src/probly/metrics/_common.py: append the three @flexdispatch entries.
- src/probly/metrics/array.py:    append numpy handlers + envelope helpers.
- src/probly/metrics/torch.py:    append torch handlers + envelope helpers.
- src/probly/metrics/__init__.py: extend the existing torch lazy-load
  callback with the three new entries; export them via __all__.
- src/probly/evaluation/__init__.py: collapse to a thin re-export.
- src/probly/evaluation/{metrics,array,torch}.py: deleted.
- src/probly/__init__.py: expose probly.metrics at the top level.
- tests/probly/evaluation/{_metrics_suite, _credal_suite, test_array,
  test_torch, test_credal, test_torch_credal}.py: git-mv'd into
  tests/probly/metrics/predicted_sets/.
- One subprocess test was checking the lazy-load module name; updated
  from probly.evaluation.torch to probly.metrics.torch.
