+++ src/probly/metrics/__init__.py
@@ -9,7 +9,10 @@ from . import array as array
 # eagerly register numpy (always available)
 from ._common import (
     auc,
+    average_interval_width,
     average_precision_score,
+    coverage,
+    efficiency,
     precision_recall_curve,
     roc_auc_score,
     roc_curve,
@@ -17,7 +20,10 @@ from ._common import (
 
 
 @auc.delayed_register((TORCH_TENSOR, TORCH_TENSOR_LIKE))
+@average_interval_width.delayed_register((TORCH_TENSOR, TORCH_TENSOR_LIKE))
 @average_precision_score.delayed_register((TORCH_TENSOR, TORCH_TENSOR_LIKE))
+@coverage.delayed_register((TORCH_TENSOR, TORCH_TENSOR_LIKE))
+@efficiency.delayed_register((TORCH_TENSOR, TORCH_TENSOR_LIKE))
 @precision_recall_curve.delayed_register((TORCH_TENSOR, TORCH_TENSOR_LIKE))
 @roc_auc_score.delayed_register((TORCH_TENSOR, TORCH_TENSOR_LIKE))
 @roc_curve.delayed_register((TORCH_TENSOR, TORCH_TENSOR_LIKE))
@@ -36,7 +42,10 @@ def _(_: type) -> None:
 
 __all__ = [
     "auc",
+    "average_interval_width",
     "average_precision_score",
+    "coverage",
+    "efficiency",
     "precision_recall_curve",
     "roc_auc_score",
     "roc_curve",
